+++ dart/geobase/test/advices/coordinates_position.dart
@@ -708,6 +708,29 @@ void _doTestPositionSeries(PositionSeries series) {
         10.0, 20.0, 30.0, 40.0,
         //
       ]);
+      expect(series.subseries(0, 0).values, const <double>[]);
+      expect(series.subseries(3, 5).values, const <double>[]);
+      expect(series.subseries(0, 1).values, [
+        10.0, 20.0, 30.0, 40.0,
+        //
+      ]);
+      expect(series.subseries(1, 2).values, [
+        12.5, 22.5, 32.5, 42.5,
+        //
+      ]);
+      expect(series.subseries(1, 10).values, [
+        12.5, 22.5, 32.5, 42.5,
+        15.0, 25.0, 35.0, 45.0,
+        //
+      ]);
+      expect(series.subseries(2).values, [
+        15.0, 25.0, 35.0, 45.0,
+        //
+      ]);
+      expect(series.reversed().subseries(2).values, [
+        10.0, 20.0, 30.0, 40.0,
+        //
+      ]);
       expect(series.project(AddOneOnXYProjection()).values, [
         11.0, 21.0, 30.0, 40.0,
         13.5, 23.5, 32.5, 42.5,
@@ -814,6 +837,10 @@ void _doTestPositionSeries(PositionSeries series) {
         10.0, 20.0, 30.0,
         //
       ]);
+      expect(series.subseries(1, 2).values, [
+        12.5, 22.5, 32.5,
+        //
+      ]);
       expect(series.project(AddOneOnXYProjection()).values, [
         11.0, 21.0, 30.0,
         13.5, 23.5, 32.5,
@@ -916,6 +943,10 @@ void _doTestPositionSeries(PositionSeries series) {
         10.0, 20.0, 40.0,
         //
       ]);
+      expect(series.subseries(1, 2).values, [
+        12.5, 22.5, 42.5,
+        //
+      ]);
       expect(series.project(AddOneOnXYProjection()).values, [
         11.0, 21.0, 40.0,
         13.5, 23.5, 42.5,
@@ -1007,6 +1038,10 @@ void _doTestPositionSeries(PositionSeries series) {
         15.0, 25.0,
         12.5, 22.5,
         10.0, 20.0,
+        //
+      ]);
+      expect(series.subseries(1, 2).values, [
+        12.5, 22.5,
         //
       ]);
       expect(series.project(AddOneOnXYProjection()).values, [